merge assaf/vanity#master

+++ test/test_helper.rb
@@ -1,11 +1,10 @@
-GC.disable
 $LOAD_PATH.delete_if { |path| path[/gems\/vanity-\d/] }
 $LOAD_PATH.unshift File.expand_path("../lib", File.dirname(__FILE__))
 ENV["RACK_ENV"] = "test"
 ENV["DB"] ||= "redis"
 
-require "test/unit"
-require "mocha"
+require "minitest/autorun"
+require "tmpdir"
 require "action_controller"
 require "action_controller/test_case"
 require "action_view/test_case"
@@ -16,73 +15,65 @@ begin
 rescue LoadError
 end
 
-if defined?(Rails::Railtie)
-  require File.expand_path("../dummy/config/environment.rb",  __FILE__)
-  require "rails/test_help"
-else
-  RAILS_ROOT = File.expand_path("..")
-  require "initializer"
-  Rails.configuration = Rails::Configuration.new
+require File.expand_path("../dummy/config/environment.rb",  __FILE__)
+require "rails/test_help"
 
-  ActionController::Routing::Routes.draw do |map|
-    map.connect ':controller/:action/:id'
-  end
-  require "phusion_passenger/events"
-end
-
-require "lib/vanity"
+require "vanity"
 require "timecop"
-require "webmock/test_unit"
 
-#Do to load order differences in Rails boot and test requires we have to manually
-#require these
+if defined?(Mocha::VERSION) && Mocha::VERSION < "0.13.0"
+  require "mocha"
+else
+  require "mocha/mini_test"
+end
+require "webmock/minitest"
+
+# Due to load order differences in Rails boot and test requires we have to
+# manually require these
+
 require 'vanity/frameworks/rails'
 Vanity::Rails.load!
 
 if $VERBOSE
   $logger = Logger.new(STDOUT)
   $logger.level = Logger::DEBUG
+else
+  $logger = Logger.new(STDOUT)
+  $logger.level = Logger::FATAL
 end
 
-
-class Test::Unit::TestCase
-  include WebMock::API
-
+module VanityTestHelpers
   # We go destructive on the database at the end of each run, so make sure we
   # don't use databases you care about. For Redis, we pick database 15
   # (default is 0).
   DATABASE = {
     "redis"=>"redis://localhost/15",
     "mongodb"=>"mongodb://localhost/vanity",
-    "mysql"=> { "adapter"=>"active_record", "active_record_adapter"=>"mysql", "database"=>"vanity_test" },
-    "postgres"=> { "adapter"=>"active_record", "active_record_adapter"=>"postgresql", "database"=>"vanity_test", "username"=>"postgres" },
+    "active_record"=> { "adapter"=>"active_record", "active_record_adapter"=>"default" },
     "mock"=>"mock:/"
   }[ENV["DB"]] or raise "No support yet for #{ENV["DB"]}"
 
-  def rails3?
-    defined?(Rails::Railtie)
-  end
-
-  def setup
+  def setup_after
     FileUtils.mkpath "tmp/experiments/metrics"
     new_playground
   end
 
-  # Call this on teardown. It wipes put the playground and any state held in it
-  # (mostly experiments), resets vanity ID, and clears database of all experiments.
-  def nuke_playground
-    Vanity.playground.connection.flushdb
-    new_playground
+  def teardown_after
+    Vanity.context = nil
+    FileUtils.rm_rf "tmp"
+    Vanity.playground.connection.flushdb if Vanity.playground.connected?
+    WebMock.reset!
   end
 
   # Call this if you need a new playground, e.g. to re-define the same experiment,
   # or reload an experiment (saved by the previous playground).
   def new_playground
     Vanity.playground = Vanity::Playground.new(:logger=>$logger, :load_path=>"tmp/experiments")
-    Vanity.playground.establish_connection DATABASE
+    ActiveRecord::Base.establish_connection
+    Vanity.playground.establish_connection(DATABASE)
   end
 
-  # Defines the specified metrics (one or more names).  Returns metric, or array
+  # Defines the specified metrics (one or more names). Returns metric, or array
   # of metric (if more than one argument).
   def metric(*names)
     metrics = names.map do |name|
@@ -115,53 +106,66 @@ class Test::Unit::TestCase
     Vanity.playground.stubs(:connection).returns(stub(:flushdb=>nil))
   end
 
-  def teardown
-    Vanity.context = nil
-    FileUtils.rm_rf "tmp"
-    Vanity.playground.connection.flushdb if Vanity.playground.connected?
-    WebMock.reset!
+  def dummy_request
+    ActionDispatch::TestRequest.new()
   end
 
+  # Defining setup/tear down in a module and including it below doesn't
+  # override the built-in setup/teardown methods, so we alias_method_chain
+  # them to run.
+  def self.included(klass)
+    klass.class_eval {
+      alias :teardown_before :teardown
+      alias :teardown :teardown_after
+
+      alias :setup_before :setup
+      alias :setup :setup_after
+    }
+  end
 end
 
-
-if  ENV["DB"] == "postgres"
-  ActiveRecord::Base.establish_connection :adapter=>"postgresql", :database=>"vanity_test"
-else
-  ActiveRecord::Base.establish_connection :adapter=>"mysql", :database=>"vanity_test"
+if defined?(Test)
+  class Test::Unit::TestCase
+    include VanityTestHelpers
+  end
 end
-ActiveRecord::Base.logger = $logger
 
-if ENV["DB"] == "mysql" || ENV["DB"] == "postgres"
+class MiniTest::Spec
+  include VanityTestHelpers
+end
+
+if defined?(MiniTest::Unit::TestCase)
+  class MiniTest::Unit::TestCase
+    include VanityTestHelpers
+  end
+end
+
+if defined?(ActiveSupport::TestCase)
+  class ActiveSupport::TestCase
+    include VanityTestHelpers
+
+    self.use_instantiated_fixtures = false if respond_to?(:use_instantiated_fixtures)
+    self.use_transactional_fixtures = false if respond_to?(:use_transactional_fixtures)
+  end
+end
+
+if defined?(ActionController::TestCase)
+  class ActionController::TestCase
+    alias :setup_controller_request_and_response_without_vanity :setup_controller_request_and_response
+    # Sets Vanity.context to the current controller, so you can do things like:
+    #   experiment(:simple).chooses(:green)
+    def setup_controller_request_and_response
+      setup_controller_request_and_response_without_vanity
+      Vanity.context = @controller
+    end
+  end
+end
+
+if ENV["DB"] == "active_record"
+  ActiveRecord::Base.establish_connection
+  ActiveRecord::Base.logger = $logger
+
   require "generators/templates/vanity_migration"
   VanityMigration.down rescue nil
   VanityMigration.up
-end
-
-
-class Array
-  # Not in Ruby 1.8.6.
-  unless method_defined?(:shuffle)
-    def shuffle
-      copy = clone
-      Array.new(size) { copy.delete_at(Kernel.rand(copy.size)) }
-    end
-  end
-end
-
-
-# Source: http://gist.github.com/25455
-def context(*args, &block)
-  return super unless (name = args.first) && block
-  parent = Class === self ? self : (defined?(ActiveSupport::TestCase) ? ActiveSupport::TestCase : Test::Unit::TestCase)
-  klass = Class.new(parent) do
-    def self.test(name, &block)
-      define_method("test_#{name.gsub(/\W/,'_')}", &block) if block
-    end
-    def self.xtest(*args) end
-    def self.setup(&block) define_method(:setup) { super() ; instance_eval &block } end
-    def self.teardown(&block) define_method(:teardown) { super() ; instance_eval &block } end
-  end
-  parent.const_set name.split(/\W+/).map(&:capitalize).join, klass
-  klass.class_eval &block
 end
